doc: remove the non-root-install mess
This was an abomination. Let's leave the past behind.

This page will be further edited when swapping out httplib2 for requests
and adding the PostgreSQL psycopg2 dependency. I'm probably going to add
a requirements.txt file for convenience reasons at that point.

When all of that is done, and we're at the point where we can make the
Debian Bullseye m2ee-tools and python3-m2ee packages, the middle part
will be updated as well.